refactor: organize nav components more cleanly
