+++ playground/src/components/Header.tsx
@@ -1,0 +1,91 @@
+import { useState } from "react";
+import { useNavigate } from 'react-router-dom';
+import MidenLogo from "./MidenLogo";
+import { Dialog } from '@headlessui/react'
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline'
+
+
+/** External links for the top navigation menu */
+const navigation = [
+    { name: 'Documentation', href: 'https://wiki.polygon.technology/docs/miden/user_docs/assembly/main' },
+    { name: 'Examples', href: 'https://github.com/0xPolygonMiden/examples#available-examples' },
+    { name: 'Homepage', href: 'https://polygon.technology/solutions/polygon-miden/' },
+]
+
+function classNames(...classes: any[]) {
+    return classes.filter(Boolean).join(' ')
+}
+
+const Header = () => {
+
+    const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
+
+    return (
+        <header className="bg-gray-900">
+            <nav className="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8" aria-label="Global">
+                <div className="flex items-center">
+                    <a href="https://polygon.technology/solutions/polygon-miden/" className="-m-1.5 px-1.5">
+                        <span className="sr-only">Polygon Miden</span>
+                        <MidenLogo className="fill-white h-10 w-auto" />
+                    </a>
+                    <h1 className="text-xl font-semibold leading-6 text-white">MidenVM Playground</h1>
+                </div>
+                <div className="flex lg:hidden">
+                    <button
+                        type="button"
+                        className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-white"
+                        onClick={() => setMobileMenuOpen(true)}
+                    >
+                        <span className="sr-only">Open main menu</span>
+                        <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+                    </button>
+                </div>
+                <div className="hidden lg:flex lg:gap-x-12">
+                    {navigation.map((item) => (
+                        <a key={item.name} href={item.href} target="_blank" rel="noopener noreferrer" className="text-sm font-semibold leading-6 text-white">
+                            {item.name}
+                        </a>
+                    ))}
+                </div>
+            </nav>
+            <Dialog as="div" className="lg:hidden" open={mobileMenuOpen} onClose={setMobileMenuOpen}>
+                <div className="fixed inset-0 z-10" />
+                <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
+                    <div className="flex items-center justify-between">
+                        <a href="#" className="-m-1.5 p-1.5">
+                            <span className="sr-only">Polygon Miden</span>
+                            <MidenLogo className="fill-gray-900 h-10 w-auto" />
+                        </a>
+                        <button
+                            type="button"
+                            className="-m-2.5 rounded-md p-2.5 text-gray-700"
+                            onClick={() => setMobileMenuOpen(false)}
+                        >
+                            <span className="sr-only">Close menu</span>
+                            <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+                        </button>
+                    </div>
+                    <div className="mt-6 flow-root">
+                        <div className="-my-6 divide-y divide-gray-500/10">
+                            <div className="space-y-2 py-6">
+                                {navigation.map((item) => (
+                                    <a
+                                        key={item.name}
+                                        href={item.href}
+                                        target="_blank"
+                                        rel="noopener noreferrer"
+                                        className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
+                                    >
+                                        {item.name}
+                                    </a>
+                                ))}
+                            </div>
+                        </div>
+                    </div>
+                </Dialog.Panel>
+            </Dialog>
+        </header>
+    );
+};
+
+export default Header;
